refactor: move acm filter and refresh logic out of cli into core
