+++ scripts/acm-permissions/cli.ts
@@ -4,22 +4,14 @@ import * as fs from "fs";
 import { parseArgs } from "node:util";
 import * as path from "path";
 
-import {
-  GUARDIANS,
-  REGISTRY_DIR,
-  STARTING_BLOCKS,
-  TIMELOCK_NAMES,
-  acmAddress,
-  filtersDir,
-  isLegacyAcm,
-  rpcUrl,
-  snapshotDir,
-} from "./config";
+import { REGISTRY_DIR, STARTING_BLOCKS, acmAddress, filtersDir, isLegacyAcm, rpcUrl, snapshotDir } from "./config";
 import { buildHashTable } from "./core/decoder";
 import { diffSnapshots } from "./core/diff";
 import { scanRange } from "./core/fetcher";
+import { expandAliases, filterPermissions, renderFilterMd, slugify } from "./core/filter";
 import { writePermissionsOutputs } from "./core/output";
 import { applyEvents } from "./core/reducer";
+import { refreshNetwork } from "./core/refresh";
 import {
   loadKnownAddresses,
   loadLegacyOnlySignatures,
@@ -243,77 +235,6 @@ function resolveNetworks(arg: string): Network[] {
   return [...new Set(requested)] as Network[];
 }
 
-export type FilterResult = Record<string, Array<{ contract: string; functionSig: string | null; roleHash: string }>>;
-
-// Expands grantee aliases BEFORE resolution, so the report keeps one section per real grantee
-// (per-timelock / per-guardian attribution survives); the output filename is built from the
-// short alias the user typed, not the expansion.
-//   - "Timelocks" → the three timelock labels
-//   - "Guardian" on a multi-guardian network → "Guardian 1".."Guardian N" (matching the
-//     registry's naming); on a single-guardian network it stays "Guardian" (the registry name)
-export const expandAliases = (labels: string[], network: Network): string[] =>
-  labels.flatMap(l => {
-    if (l === "Timelocks") return TIMELOCK_NAMES;
-    if (l === "Guardian" && GUARDIANS[network].length > 1) return GUARDIANS[network].map((_, i) => `Guardian ${i + 1}`);
-    return [l];
-  });
-
-// Resolves a requested grantee label to the address(es) it stands for:
-//   - a raw `0x…` address resolves to itself (checksummed)
-//   - the literal "Guardian" resolves to ALL guardian multisigs on the network
-//   - anything else is looked up in the name map (address whose mapped name matches the label;
-//     registries join conflicting deployments as "A / B" — e.g. "X / X_Proxy" — so each
-//     " / "-separated part matches on its own)
-// Throws a clear error when a label resolves to nothing, rather than silently matching zero
-// permissions (which would look identical to "this grantee legitimately holds nothing").
-function resolveLabel(label: string, network: Network, nameMap: Record<string, string>): string[] {
-  if (label.startsWith("0x")) return [ethers.utils.getAddress(label)];
-  if (label === "Guardian") return GUARDIANS[network];
-  const addresses = Object.keys(nameMap).filter(addr => nameMap[addr].split(" / ").includes(label));
-  if (addresses.length === 0) {
-    throw new Error(
-      `unknown grantee label "${label}" — expected a 0x… address, "Guardian", "Timelocks", or a name ` +
-        `present in the ${network} contract registry`,
-    );
-  }
-  return addresses;
-}
-
-export function filterPermissions(
-  file: SnapshotFile,
-  grantees: string[],
-  network: Network,
-  nameMap: Record<string, string> = loadNameMap(network),
-  exclude: string[] = [],
-  onlySigs: Set<string> | null = null,
-): FilterResult {
-  const result: FilterResult = {};
-  // Set difference on the exact permission entry (same contract + roleHash): a permission is
-  // dropped only when an excluded grantee holds that very role, not merely the same function
-  // signature on a different contract.
-  const excluded = new Set(
-    exclude.flatMap(label => resolveLabel(label, network, nameMap)).map(a => ethers.utils.getAddress(a)),
-  );
-
-  for (const label of grantees) {
-    const targets = new Set(resolveLabel(label, network, nameMap).map(a => ethers.utils.getAddress(a)));
-    const matches: FilterResult[string] = [];
-    for (const contract of file.contracts) {
-      for (const permission of contract.permissions) {
-        if (onlySigs && (permission.functionSig === null || !onlySigs.has(permission.functionSig))) continue;
-        const isGrantedToLabel = permission.grantees.some(g => targets.has(ethers.utils.getAddress(g.address)));
-        const isAlsoExcluded = permission.grantees.some(g => excluded.has(ethers.utils.getAddress(g.address)));
-        if (isGrantedToLabel && !isAlsoExcluded) {
-          matches.push({ contract: contract.name, functionSig: permission.functionSig, roleHash: permission.roleHash });
-        }
-      }
-    }
-    result[label] = matches;
-  }
-
-  return result;
-}
-
 function resolveSingleNetwork(arg: string): Network {
   if (arg === "all" || arg.includes(",")) {
     throw new Error(`filter requires exactly one --network (got "${arg}") — "all" and lists are not supported`);
@@ -322,43 +243,6 @@ function resolveSingleNetwork(arg: string): Network {
     throw new Error(`unknown network: ${arg} — expected one of ${NETWORKS.join(", ")}`);
   }
   return arg as Network;
-}
-
-// "NormalTimelock, 0xAbC…" -> "NormalTimelock+0xAbC…" with anything path-hostile replaced.
-const slugify = (labels: string[]) => labels.map(l => l.replace(/[^A-Za-z0-9_.-]+/g, "-")).join("+");
-
-function renderFilterMd(
-  meta: {
-    network: Network;
-    height: number;
-    updatedAt: string;
-    grantees: string[];
-    exclude: string[];
-    legacyOnly?: boolean;
-  },
-  result: FilterResult,
-): string {
-  const lines: string[] = [
-    `# Permission filter — ${meta.network}`,
-    "",
-    `- Grantees: ${meta.grantees.join(", ")}`,
-    ...(meta.exclude.length ? [`- Excluding permissions also held by: ${meta.exclude.join(", ")}`] : []),
-    ...(meta.legacyOnly
-      ? ["- Only permissions whose signature exists solely in legacy-signatures.json (no package source proves it)"]
-      : []),
-    `- Snapshot height: ${meta.height} (updated ${meta.updatedAt})`,
-    "",
-  ];
-  for (const label of meta.grantees) {
-    const matches = result[label];
-    lines.push(`## ${label} — ${matches.length} permission${matches.length === 1 ? "" : "s"}`, "");
-    if (matches.length) {
-      lines.push("| Contract | Function | Role hash |", "| --- | --- | --- |");
-      for (const m of matches) lines.push(`| ${m.contract} | ${m.functionSig ?? "(undecoded)"} | ${m.roleHash} |`);
-      lines.push("");
-    }
-  }
-  return lines.join("\n");
 }
 
 function filterCommand(values: {
@@ -484,51 +368,6 @@ async function fetchCommand(values: {
   if (anyFailed) process.exit(1);
 }
 
-export interface RefreshResult {
-  newlyDecoded: number;
-  total: number;
-  unresolved: number;
-}
-
-// Offline re-annotation + re-render of a committed snapshot — no RPC calls, no height change. Only
-// bscmainnet can ever produce newlyDecoded > 0, since only its roles are ever undecoded (legacy
-// hash-based roles awaiting registry growth); other networks decode fully at fetch time.
-// Returns null (after warning) when there is no snapshot to refresh for `network`.
-export function refreshNetwork(network: Network, opts: { baseDir?: string } = {}): RefreshResult | null {
-  const { baseDir } = opts;
-  const file = loadSnapshotFile(network, baseDir);
-  if (!file) {
-    console.warn(`[${network}] no snapshot — skipping`);
-    return null;
-  }
-
-  const state = fileToState(file);
-  const table = isLegacyAcm(network) ? buildHashTable(loadKnownAddresses(network), loadSignatures()) : null;
-  const before = Object.values(state).filter(r => !r.decoded).length;
-  reannotateUndecoded(state, table);
-  const after = Object.values(state).filter(r => !r.decoded).length;
-
-  const names = loadNameMap(network);
-  const newFile = stateToFile(
-    state,
-    {
-      network: file.network,
-      acmAddress: file.acmAddress,
-      height: file.height,
-      updatedAt: file.updatedAt,
-      // refresh makes no chain calls, so it cannot change the verification status — carry the
-      // loaded file's verified/verifiedAt through unchanged.
-      verified: file.verified,
-      verifiedAt: file.verifiedAt,
-    },
-    names,
-  );
-  saveSnapshotFile(network, newFile, baseDir);
-  writePermissionsOutputs(network, newFile, names, baseDir);
-
-  return { newlyDecoded: before - after, total: Object.keys(state).length, unresolved: after };
-}
-
 function refreshCommand(values: { network?: string }): void {
   let selected: Network[];
   try {
@@ -557,8 +396,8 @@ function refreshCommand(values: { network?: string }): void {
   if (anyFailed) process.exit(1);
 }
 
-// Guarded so importing this module (e.g. from tests, for `filterPermissions`) never triggers
-// the CLI dispatch below — only running `cli.ts` directly does.
+// Guarded so importing this module never triggers the CLI dispatch below — only running
+// `cli.ts` directly does.
 if (require.main === module) {
   (async () => {
     const { positionals, values } = parseArgs({
